+++ app/src/main/java/budget/application/db/mapper/TransactionRowMappers.java
@@ -17,7 +17,7 @@ public class TransactionRowMappers {
           rs.getObject("txn_date", LocalDateTime.class),
           rs.getString("merchant"),
           rs.getObject("account_id", UUID.class),
-          rs.getDouble("total_amount"),
+          rs.getBigDecimal("total_amount"),
           rs.getString("notes"),
           rs.getObject("created_at", LocalDateTime.class),
           rs.getObject("updated_at", LocalDateTime.class));
@@ -33,7 +33,7 @@ public class TransactionRowMappers {
           rs.getObject("txn_id", UUID.class),
           rs.getObject("txn_date", LocalDateTime.class),
           rs.getString("txn_merchant"),
-          rs.getDouble("txn_total_amount"),
+          rs.getBigDecimal("txn_total_amount"),
           rs.getString("txn_notes"),
           null,
           List.of());
