+++ app/src/main/java/budget/application/model/dto/TransactionRequest.java
@@ -1,5 +1,6 @@
 package budget.application.model.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -8,6 +9,6 @@ public record TransactionRequest(
     LocalDateTime txnDate,
     String merchant,
     UUID accountId,
-    double totalAmount,
+    BigDecimal totalAmount,
     String notes,
     List<TransactionItemRequest> items) {}
