+++ app/src/main/java/budget/application/model/entity/TransactionItem.java
@@ -1,6 +1,12 @@
 package budget.application.model.entity;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record TransactionItem(
-    UUID id, UUID transactionId, UUID categoryId, String label, double amount, String expType) {}
+    UUID id,
+    UUID transactionId,
+    UUID categoryId,
+    String label,
+    BigDecimal amount,
+    String expType) {}
